+++ src/test/java/org/folio/dcb/listener/CirculationRequestEventListenerTest.java
@@ -36,10 +36,8 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.MessageHeaders;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
 class CirculationRequestEventListenerTest extends BaseTenantIntegrationTest {
 
   private static final String TENANT = "diku";