+++ src/test/java/org/folio/dcb/support/kafka/KafkaContainerExtension.java
@@ -1,5 +1,6 @@
 package org.folio.dcb.support.kafka;
 
+import static org.folio.dcb.utils.EntityUtils.TEST_TENANT;
 import static org.testcontainers.utility.DockerImageName.parse;
 
 import java.util.List;
@@ -45,6 +46,7 @@ public class KafkaContainerExtension implements BeforeAllCallback, AfterAllCallb
       .map(topicName -> new NewTopic(topicName, 1, (short) 1))
       .toList();
 
+    log.info("Creating topics: {}", newTopics);
     try (var adminClient = getAdminClient()) {
       adminClient.createTopics(newTopics);
     }
@@ -53,7 +55,17 @@ public class KafkaContainerExtension implements BeforeAllCallback, AfterAllCallb
   @SneakyThrows
   public static void deleteTopics(List<String> topicNames) {
     try (var adminClient = getAdminClient()) {
-      adminClient.deleteTopics(topicNames);
+      var existingTopics = adminClient.listTopics().names().get();
+      var topicsToDelete = topicNames.stream()
+        .filter(existingTopics::contains)
+        .toList();
+
+      if (!topicsToDelete.isEmpty()) {
+        log.info("Removing Kafka topics: {}", topicsToDelete);
+        adminClient.deleteTopics(topicsToDelete);
+      } else {
+        log.debug("No Kafka topics to delete from: {}", topicNames);
+      }
     }
   }